Move `Pipeline` to `miningpipeline.MiningPipeline`
